refactor repos, services and DI
