+++ src/main/kotlin/domain/model/coin/CoinModel.kt
@@ -1,5 +1,7 @@
 package domain.model.coin
 
+import network.model.coin.CoinDto
+
 data class CoinModel(
     val symbol: String,
     val name: String,
@@ -8,4 +10,18 @@ data class CoinModel(
     val delta24H: Double,
     val delta7D: Double,
     val delta30D: Double
-)
+) {
+    companion object {
+        fun fromDto(coinDto: CoinDto): CoinModel {
+            return CoinModel(
+                symbol = coinDto.symbol,
+                name = coinDto.name,
+                price = coinDto.price.take(7).toDouble(),
+                delta1H = coinDto.delta1H.toDouble(),
+                delta24H = coinDto.delta24H.toDouble(),
+                delta7D = coinDto.delta7D.toDouble(),
+                delta30D = coinDto.delta30D.toDouble()
+            )
+        }
+    }
+}
