--- src/main/kotlin/domain/model/country/CountryModel.kt
@@ -1,6 +1,0 @@
-package domain.model.country
-
-data class CountryModel(
-    val countryName: String,
-    val probability: Int
-)
