--- src/main/kotlin/domain/model/country/CountryOfOriginModel.kt
@@ -1,5 +1,0 @@
-package domain.model.country
-
-data class CountryOfOriginModel(
-    val countries: List<CountryModel>
-)
