--- src/main/kotlin/domain/model/gender/GenderModel.kt
@@ -1,6 +1,0 @@
-package domain.model.gender
-
-data class GenderModel(
-    val gender: String,
-    val accuracy: Int
-)
