--- src/main/kotlin/domain/model/guide/GuideFieldModel.kt
@@ -1,6 +1,0 @@
-package domain.model.guide
-
-data class GuideFieldModel(
-    val title: String,
-    val content: String
-)
