--- src/main/kotlin/domain/model/guide/GuideSingleJsonModel.kt
@@ -1,7 +1,0 @@
-package domain.model.guide
-
-data class GuideSingleJsonModel(
-    val title: String,
-    val description: String? = null,
-    val fields: List<GuideFieldModel>? = null
-)
