--- src/main/kotlin/domain/util/EntityMapper.kt
@@ -1,7 +1,0 @@
-package domain.util
-
-interface EntityMapper<Entity, Model> {
-
-    suspend fun mapToModel(entity: Entity): Model
-
-}
