--- src/main/kotlin/network/model/country/CountryDto.kt
@@ -1,8 +1,0 @@
-package network.model.country
-
-import com.google.gson.annotations.SerializedName
-
-data class CountryDto(
-    @SerializedName("country_name") val countryName: String,
-    val probability: Double
-)
