--- src/main/kotlin/network/model/country/CountryOfOriginDto.kt
@@ -1,8 +1,0 @@
-package network.model.country
-
-import com.google.gson.annotations.SerializedName
-
-data class CountryOfOriginDto(
-    @SerializedName("country_of_origin")
-    val countryOrigin: List<CountryDto>
-)
