--- src/main/kotlin/network/model/gender/GenderDto.kt
@@ -1,6 +1,0 @@
-package network.model.gender
-
-data class GenderDto(
-    val gender: String,
-    val accuracy: Int
-)
