+++ src/main/kotlin/network/model/genderapi/GenderDto.kt
@@ -1,0 +1,6 @@
+package network.model.genderapi
+
+data class GenderDto(
+    val gender: String,
+    val accuracy: Int
+)
