+++ src/main/kotlin/network/service/CoinlibService.kt
@@ -1,0 +1,16 @@
+package network.service
+
+import network.model.coin.CoinDto
+import retrofit2.http.GET
+import retrofit2.http.Query
+
+interface CoinlibService {
+
+    @GET("coin")
+    suspend fun get(
+        @Query("key") token: String,
+        @Query("pref") pref: String,
+        @Query("symbol") symbol: String
+    ): CoinDto
+
+}
