--- src/main/kotlin/network/service/GenderService.kt
@@ -1,15 +1,0 @@
-package network.service
-
-import network.model.gender.GenderDto
-import retrofit2.http.GET
-import retrofit2.http.Query
-
-interface GenderService {
-
-    @GET("get")
-    suspend fun get(
-        @Query("key") token: String,
-        @Query("name") name: String
-    ): GenderDto
-
-}
