--- src/main/kotlin/repository/coin/CoinRepository.kt
@@ -1,13 +1,0 @@
-package repository.coin
-
-import domain.model.coin.CoinModel
-
-interface CoinRepository {
-
-    suspend fun get(
-        token: String,
-        pref: String,
-        symbol: String
-    ): CoinModel
-
-}
