--- src/main/kotlin/repository/country/CountryRepository.kt
@@ -1,12 +1,0 @@
-package repository.country
-
-import domain.model.country.CountryOfOriginModel
-
-interface CountryRepository {
-
-    suspend fun get(
-        token: String,
-        name: String
-    ): CountryOfOriginModel
-
-}
