--- src/main/kotlin/repository/gender/GenderRepository.kt
@@ -1,12 +1,0 @@
-package repository.gender
-
-import domain.model.gender.GenderModel
-
-interface GenderRepository {
-
-    suspend fun get(
-        token: String,
-        name: String
-    ): GenderModel
-
-}
