+++ src/main/kotlin/repository/guide/GuideRepository.kt
@@ -1,6 +1,7 @@
 package repository.guide
 
 import domain.model.guide.GuideJsonModel
+import network.service.GuideService
 
 interface GuideRepository {
 
@@ -9,4 +10,18 @@ interface GuideRepository {
         language: String
     ): GuideJsonModel
 
+}
+
+class GuideRepositoryImpl(
+    private val service: GuideService,
+) : GuideRepository {
+
+    override suspend fun fetch(
+        jsonName: String,
+        language: String,
+    ): GuideJsonModel {
+        return GuideJsonModel.fromDto(
+            service.get(jsonName, language)
+        )
+    }
 }
